+++ packages/server/src/api/api.module.ts
@@ -25,7 +25,7 @@ import { CustomersController } from './customers.controller.js'
 import { DatabaseExceptionFilter } from './database-errors.js'
 import { DocumentFiles } from './document-files.js'
 import { DocumentInstructionsController } from './document-instructions.controller.js'
-import { PaymentsController } from './payments.controller.js'
+import { OpenPaymentsController, PaymentsController } from './payments.controller.js'
 import { DocumentLinesController, DocumentTotalsController } from './document-lines.controller.js'
 import { DocumentMailController } from './document-mail.controller.js'
 import { DocumentPdfController } from './document-pdf.controller.js'
@@ -192,6 +192,7 @@ export class ApiModule implements NestModule {
         DocumentMailController,
         DocumentInstructionsController,
         PaymentsController,
+        OpenPaymentsController,
         SyncController,
         SettingsController,
         ReportFieldsController,